requirements.txt
streamlit==1.33.0
google-generativeai
PyMuPDF
python-dotenv